requirements.txt
pandas==2.1.1
numpy==1.26.1
matplotlib==3.8.0
torch==2.1.0
torch-optimizer==0.3.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchsampler==0.1.2
torchsummary==1.5.1
torchvision==0.16.0
imgaug==0.4.0
scikit-image==0.20.0
scikit-learn==1.3.1
scikit-multilearn==0.2.0
streamlit==1.28.2
sentence-transformers==2.2.2
jinja2-time==0.2.0
tqdm==4.65.0
Pillow==9.4.0
seaborn==0.12.2
matplotlib-inline==0.1.6
